Move scripts inside ui
